Remove unused code and reorganize
