--- src/MxInfo.DeviceManager/MxInfo.DeviceManager.Config.Tests.Contracts/ConfigApiApplication.cs
@@ -1,16 +1,0 @@
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.VisualStudio.TestPlatform.TestHost;
-
-namespace MxInfo.DeviceManager.Config.Tests.Contracts;
-
-internal class ConfigApiApplication : WebApplicationFactory<Program>
-{
-
-    protected override void ConfigureWebHost(IWebHostBuilder builder)
-    {
-        Environment.SetEnvironmentVariable("CacheSettings:UseCache", "false");
-        
-    }
-    
-}
