--- src/MxInfo.DeviceManager/MxInfo.DeviceManager.Config.Tests.Contracts/ConfigApiStartup.cs
@@ -1,37 +1,0 @@
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-
-namespace MxInfo.DeviceManager.Config.Tests.Contracts;
-
-public class ConfigApiStartup
-{
-    public IConfiguration Configuration { get; }
-    
-    public ConfigApiStartup(IConfiguration configuration)
-    {
-        Configuration = configuration;
-    }
-    
-    public void ConfigureServices(IServiceCollection services)
-    {
-        services.AddControllers();
-        
-    }
-    
-    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-    {
-        if (env.IsDevelopment())
-        {
-            app.UseDeveloperExceptionPage();
-        }
-
-        app.UseRouting();
-        app.UseEndpoints(endpoints =>
-        {
-            endpoints.MapControllers();
-        });
-    }
-}
