+++ src/MxInfo.DeviceManager/MxInfo.DeviceManager.Config.Tests.Contracts/Infrastructure/ConfigurationApiFixtureCollection.cs
